display annotation integers as integers
